+++ changelogs/2026-06-04_page-tutorial-scope-fix.md
@@ -1,2 +1,3 @@
 | fix | prd-admin | 修复「本页教程」打开却弹出别页教程:教程抽屉改为按当前路由作用域过滤,移除随机兜底,无本页教程时给空态+「浏览全部」入口 |
 | fix | prd-admin | 修复多步引导连续找不到锚点时旧光圈滞留导致「每步都指向同一元素」的错觉(SpotlightOverlay 切步时若目标不在 DOM 先清光圈改显「正在定位」) |
+| fix | prd-admin | 本页没有任何教程时不再显示「本页教程」入口按钮;按钮显隐与抽屉作用域共用 filterPageTips 单一过滤逻辑(SSOT) |
