+++ prd-admin/src/components/daily-tips/pageGuideMatch.ts
@@ -40,3 +40,38 @@ export function matchPageGuide(
     return false;
   }) ?? null;
 }
+
+/**
+ * 当前路由「本页教程」集合 —— TipsDrawer(抽屉作用域)与 TipsEntryButton(按钮是否显示)
+ * 共用的单一过滤逻辑(SSOT,避免两份 inline 拷贝随规则漂移)。
+ *
+ * 规则:
+ * - 只看 card / spotlight 类,排除已 dismiss 的。
+ * - 管理员定向推送(isTargeted):不限页面,始终纳入(否则会漏掉用户被推送的内容)。
+ * - *-page-guide:按编辑器感知规则匹配(非编辑器走精确路由、编辑器走深层路由前缀),
+ *   与 matchPageGuide 一致,避免列表页教程在编辑器子路由里被误纳入。
+ * - 其余(功能公告 / 旧版短教程):按 actionUrl 精确或前缀匹配当前页。
+ *
+ * 入参用 store 的稳定引用 items + dismissed,这样调用方 useMemo([items, dismissed, pathname]) 能真正缓存。
+ */
+export function filterPageTips(
+  items: DailyTip[],
+  dismissed: Set<string>,
+  pathname: string,
+): DailyTip[] {
+  return items.filter((t) => {
+    if (dismissed.has(t.id)) return false;
+    if (t.kind !== 'card' && t.kind !== 'spotlight') return false;
+    if (t.isTargeted) return true;
+    if (!t.actionUrl) return false;
+    const url = t.actionUrl;
+    const isPageGuide = typeof t.sourceId === 'string' && t.sourceId.endsWith('-page-guide');
+    if (isPageGuide) {
+      const isEditor = isEditorPageGuide(t.sourceId);
+      if (pathname === url) return !isEditor;
+      if (pathname.startsWith(url + '/') || pathname.startsWith(url + '-fullscreen/')) return isEditor;
+      return false;
+    }
+    return pathname === url || pathname.startsWith(url + '/');
+  });
+}
